+++ src/app/page.tsx
@@ -2,59 +2,36 @@
 
 "use client";
 
-import React, { useEffect, useRef, useState, FormEvent } from "react";
-import { Context } from "@/components/Context";
+import React, { useEffect, useState } from "react";
+import { Sidebar } from "@/components/Sidebar";
 import Chat from "@/components/Chat";
-import InstructionModal from "./components/InstructionModal";
-import { Message } from "ai";
-import { PineconeRecord, ScoredPineconeRecord } from "@pinecone-database/pinecone";
+import AppContext from "./appContext";
+import useRefreshIndex from '@/hooks/useRefreshIndex'
+import type { PineconeRecord } from "@pinecone-database/pinecone";
 
 const Page: React.FC = () => {
   const [context, setContext] = useState<{ context: PineconeRecord[] }[] | null>(null);
-  const [isModalOpen, setModalOpen] = useState(false);
+  const { totalRecords, refreshIndex } = useRefreshIndex();
 
-  const [totalRecords, setTotalRecords] = useState<number>(0);
-
-
-  const refreshIndex = async () => {
-    const response = await fetch("/api/checkIndex", {
-      method: "POST",
-    });
-    try {
-      const stats = await response.json();
-      setTotalRecords(stats.totalRecordCount);
-    } catch (e) {
-      console.log(e)
+  useEffect(() => {
+    if (totalRecords === 0) {
+      refreshIndex()
     }
-  }
-
+  }, [refreshIndex, totalRecords])
 
   return (
-    <div className="flex flex-col justify-between h-screen bg-whitemx-auto max-w-full">
-      <InstructionModal
-        isOpen={isModalOpen}
-        onClose={() => setModalOpen(false)}
-      />
-      <div className="flex w-full flex-grow overflow-hidden relative">
-        <div style={{
-          backgroundColor: "#FBFBFC"
-        }} className="absolute transform translate-x-full transition-transform duration-500 ease-in-out right-0 w-2/3 h-full bg-white overflow-y-auto lg:static lg:translate-x-0 lg:w-2/5">
-          <Context className="" context={context} refreshIndex={refreshIndex} />
+    <AppContext.Provider value={{ totalRecords, refreshIndex }}>
+      <div className="flex flex-col justify-between h-screen bg-whitemx-auto max-w-full">
+        <div className="flex w-full flex-grow overflow-hidden relative">
+          <div style={{
+            backgroundColor: "#FBFBFC"
+          }} className="absolute transform translate-x-full transition-transform duration-500 ease-in-out right-0 w-2/3 h-full bg-white overflow-y-auto lg:static lg:translate-x-0 lg:w-2/5">
+            <Sidebar />
+          </div>
+          <Chat setContext={setContext} context={context} />
         </div>
-        <Chat setContext={setContext} showIndexMessage={totalRecords === 0} context={context} />
-        <button
-          type="button"
-          className="absolute left-20 transform -translate-x-12 bg-white text-white rounded-l py-2 px-4 lg:hidden"
-          onClick={(e) => {
-            e.currentTarget.parentElement
-              ?.querySelector(".transform")
-              ?.classList.toggle("translate-x-full");
-          }}
-        >
-          ☰
-        </button>
       </div>
-    </div>
+    </AppContext.Provider>
   );
 };
