+++ src/app/utils/svg/pinecone.tsx
@@ -1,5 +1,5 @@
 export const PineconeSvg = () => {
-    return <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+    return <svg width="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
         <g id="Group 1">
             <path id="Union" fillRule="evenodd" clipRule="evenodd" d="M9.86515 0.436421C9.71202 0.254145 9.45051 0.207319 9.24364 0.325131L9.05019 0.435301L9.04787 0.434873L9.04751 0.436828L6.96586 1.62232L7.45386 2.47922L8.81343 1.70494L8.47896 3.51696L9.44869 3.69596L9.78499 1.87402L10.7864 3.06604L11.5415 2.43172L10.017 0.617155L10.0176 0.613871L10.0136 0.613136L9.86515 0.436421ZM6.93092 15.7395C7.43237 15.7395 7.83887 15.3433 7.83887 14.8545C7.83887 14.3657 7.43237 13.9694 6.93092 13.9694C6.42946 13.9694 6.02296 14.3657 6.02296 14.8545C6.02296 15.3433 6.42946 15.7395 6.93092 15.7395ZM8.152 10.8826L7.82491 12.7064L6.85429 12.5323L7.17933 10.7199L5.82444 11.498L5.33339 10.6428L7.40668 9.45227L7.40713 9.44973L7.41015 9.45027L7.60473 9.33854C7.81085 9.22018 8.07207 9.26572 8.22598 9.44684L8.37611 9.62351L8.37776 9.62381L8.37751 9.62516L9.91555 11.4351L9.16411 12.0736L8.152 10.8826ZM8.93747 6.52891L8.61064 8.35141L7.64002 8.17735L7.9637 6.37237L6.61348 7.14381L6.12429 6.28759L8.19061 5.10702L8.19285 5.09455L8.20775 5.09723L8.39616 4.98958C8.60225 4.87183 8.86298 4.91763 9.0166 5.09856L9.16055 5.26809L9.16348 5.26861L9.16304 5.27102L10.7005 7.08168L9.94876 7.71993L8.93747 6.52891ZM2.57273 13.725L2.56741 13.7286L2.56308 13.7223L2.36224 13.6646C2.14946 13.6035 2.0086 13.4017 2.02465 13.1809L2.2034 10.7216L3.13691 10.7894L3.02819 12.2852L4.48216 11.2991L5.00753 12.0737L3.58144 13.0409L5.03669 13.4588L4.77835 14.3584L2.57273 13.725ZM11.4482 15.2791L11.4556 15.2896L11.4685 15.2804L11.6739 15.294C11.891 15.3085 12.0894 15.1714 12.1526 14.9632L12.8681 12.6098L11.9726 12.3376L11.533 13.7835L10.5172 12.3555L9.7545 12.898L10.7479 14.2945L9.22464 14.1934L9.16263 15.1273L11.4482 15.2791ZM14.8974 9.6713L14.8998 9.67173L14.8993 9.6747L15.0083 9.87004C15.1199 10.07 15.0745 10.321 14.8998 10.469L14.7349 10.6088L14.7337 10.6155L14.7282 10.6145L12.9523 12.1199L12.3327 11.3889L13.501 10.3986L11.7513 10.0907L11.9173 9.1469L13.6804 9.45715L12.941 8.13252L13.7777 7.66546L14.8974 9.6713ZM13.096 5.55754L11.5151 6.40817L11.0611 5.56432L12.6144 4.7285L11.1868 4.15106L11.5462 3.26272L13.711 4.13841L13.7208 4.13312L13.7272 4.14495L13.9232 4.22423C14.1336 4.30932 14.2554 4.53031 14.215 4.75361L14.1747 4.97662L14.1749 4.97697L14.1746 4.97712L13.7678 7.22857L12.8248 7.05815L13.096 5.55754ZM2.49916 7.5346L4.25266 7.83297L4.09191 8.77764L2.32373 8.47678L3.07777 9.8054L2.24437 10.2784L1.10414 8.26925L1.10295 8.26905L1.10319 8.26759L0.994093 8.07535C0.881082 7.87622 0.924752 7.62496 1.09832 7.47564L1.26189 7.33493L1.26369 7.32437L1.27243 7.32586L3.02679 5.81664L3.65172 6.54308L2.49916 7.5346ZM5.14808 3.44856L6.31515 4.74872L5.60205 5.38883L4.40808 4.05871L4.13015 5.5735L3.18763 5.40057L3.59918 3.15756L3.59722 3.15539L3.60004 3.15286L3.64174 2.92561C3.68212 2.7055 3.8695 2.54266 4.09309 2.53336L4.29976 2.52476L4.31032 2.51528L4.31815 2.524L6.66067 2.42657L6.70049 3.384L5.14808 3.44856Z" fill="#030080" />
         </g>
