+++ src/app/utils/svg/pineconeLogo.tsx
@@ -1,5 +1,5 @@
 export const PineconeLogoSvg = () => {
-    return (<svg xmlns="http://www.w3.org/2000/svg" width="127" height="25" viewBox="0 0 127 25" fill="none">
+    return (<svg xmlns="http://www.w3.org/2000/svg" width="127" viewBox="0 0 127 25" fill="none">
         <path d="M28.5078 5.81689H35.0289C39.7481 5.81689 40.9434 8.59076 40.9434 10.8516C40.9434 13.1124 39.7213 15.8863 35.0289 15.8863H31.0056V23.8166H28.5078V5.81689ZM31.0056 13.7472H34.2974C36.2779 13.7472 38.2583 13.295 38.2583 10.8516C38.2583 8.40816 36.2779 7.95599 34.2974 7.95599H31.0056V13.7472Z" fill="black" />
         <path d="M44.2089 5.73016C45.1515 5.73077 45.9184 6.47011 45.9293 7.38886C45.9403 8.30762 45.1913 9.06418 44.249 9.0862C43.3067 9.10821 42.5213 8.3875 42.4872 7.46929C42.4727 7.01442 42.6474 6.57315 42.9716 6.24567C43.2958 5.9182 43.742 5.73229 44.2089 5.73016ZM43.0581 11.7649H45.3775V23.8171H43.0581V11.7649Z" fill="black" />
         <path d="M48.0615 11.7652H50.4077V13.626H50.4612C51.2792 12.2038 52.8594 11.3627 54.5291 11.4609C56.9288 11.4609 58.9895 12.8608 58.9895 16.0607V23.817H56.6701V16.7042C56.6701 14.4347 55.3409 13.5999 53.8511 13.5999C51.8975 13.5999 50.4077 14.8173 50.4077 17.6172V23.817H48.0615V11.7652Z" fill="black" />
@@ -8,6 +8,6 @@ export const PineconeLogoSvg = () => {
         <path d="M92.4784 11.4609C96.0506 11.5183 98.9044 14.3774 98.8653 17.8595C98.8262 21.3416 95.9089 24.1391 92.3362 24.1204C88.7636 24.1014 85.8774 21.2734 85.877 17.7911C85.8767 16.094 86.5756 14.4679 87.8168 13.2777C89.058 12.0875 90.7375 11.433 92.4784 11.4609ZM92.4784 21.9909C94.9762 21.9909 96.5463 20.2518 96.5463 17.7911C96.5463 15.3303 94.9762 13.6 92.4784 13.6C89.9805 13.6 88.4015 15.339 88.4015 17.7911C88.4015 20.2431 89.9716 21.9909 92.4784 21.9909Z" fill="black" />
         <path d="M100.954 11.765H103.3V13.6258C104.121 12.2022 105.705 11.3612 107.377 11.4606C109.777 11.4606 111.838 12.8606 111.838 16.0605V23.8169H109.491V16.704C109.491 14.4345 108.162 13.5997 106.681 13.5997C104.719 13.5997 103.238 14.8171 103.238 17.617V23.8169H100.954V11.765Z" fill="black" />
         <path d="M116.208 18.6343C116.208 20.7386 118.215 22.1126 120.383 22.1126C121.741 22.0699 122.999 21.4021 123.773 20.3126L125.557 21.6343C124.26 23.2862 122.21 24.2072 120.079 24.0952C116.172 24.0952 113.719 21.356 113.719 17.7647C113.649 16.0842 114.296 14.4505 115.506 13.2513C116.716 12.0522 118.381 11.3945 120.106 11.4343C124.469 11.4343 126.137 14.6865 126.137 17.7908V18.6082L116.208 18.6343ZM123.612 16.7995C123.559 14.7908 122.417 13.3212 120.044 13.3212C118.02 13.3034 116.336 14.8309 116.208 16.7995H123.612Z" fill="black" />
-        <path fill-rule="evenodd" clip-rule="evenodd" d="M14.402 0.492068C14.1568 0.200153 13.738 0.125161 13.4067 0.313837L13.0967 0.490347L13.0926 0.489595L13.092 0.493029L9.75878 2.39129L10.5403 3.76361L12.7171 2.52392L12.1815 5.42557L13.7345 5.71223L14.2732 2.79391L15.8774 4.7034L17.0866 3.68754L14.6448 0.781041L14.6457 0.776261L14.6399 0.775191L14.402 0.492068ZM9.70257 25C10.5056 25 11.1567 24.3654 11.1567 23.5826C11.1567 22.7998 10.5056 22.1653 9.70257 22.1653C8.8995 22.1653 8.24847 22.7998 8.24847 23.5826C8.24847 24.3654 8.8995 25 9.70257 25ZM11.6579 17.2218L11.1341 20.1425L9.57964 19.8637L10.1002 16.9611L7.9302 18.2072L7.14378 16.8376L10.4643 14.9309L10.465 14.9269L10.4698 14.9277L10.7813 14.7489C11.1114 14.5593 11.5298 14.6322 11.7763 14.9223L12.0166 15.2051L12.0195 15.2057L12.0191 15.208L14.4821 18.1065L13.2787 19.1291L11.6579 17.2218ZM12.9159 10.2489L12.3924 13.168L10.838 12.8892L11.3563 9.99873L9.1943 11.234L8.41086 9.86274L11.7197 7.97226L11.7233 7.95209L11.7475 7.95641L12.0493 7.78397C12.3793 7.59541 12.7969 7.66875 13.0429 7.95851L13.2735 8.23008L13.2778 8.23085L13.2772 8.2344L15.7396 11.1345L14.5358 12.1566L12.9159 10.2489ZM2.72286 21.7738L2.71439 21.7796L2.70749 21.7694L2.38579 21.677C2.04503 21.5792 1.81944 21.2559 1.84514 20.9023L2.13142 16.9638L3.62643 17.0724L3.45232 19.4679L5.78088 17.8886L6.62225 19.1291L4.33831 20.6782L6.66893 21.3474L6.25521 22.7882L2.72286 21.7738ZM18.1706 23.4104L18.1707 23.4106L18.1704 23.4108L18.0653 23.7567C17.9639 24.0901 17.6462 24.3097 17.2985 24.2866L16.9699 24.2648L16.9492 24.2794L16.9373 24.2626L13.2768 24.0196L13.3761 22.5239L15.8157 22.6859L14.2249 20.4493L15.4464 19.5805L17.073 21.8673L17.7769 19.5518L19.211 19.9878L18.1706 23.4104ZM22.4612 15.2819L22.4648 15.2825L22.4641 15.2869L22.6388 15.6C22.8176 15.9202 22.7447 16.3222 22.465 16.5593L22.2007 16.7833L22.1989 16.7939L22.1901 16.7924L19.3462 19.2031L18.3538 18.0325L20.2248 16.4465L17.4224 15.9534L17.6884 14.442L20.5122 14.9389L19.328 12.8174L20.668 12.0694L22.4612 15.2819ZM19.5763 8.6934L17.0442 10.0559L16.317 8.70445L18.8048 7.36579L16.5187 6.44108L17.0942 5.01841L20.5611 6.42075L20.5766 6.41239L20.5867 6.43112L20.9011 6.55827C21.2379 6.69454 21.433 7.04845 21.3684 7.40606L21.3038 7.76371L21.3038 7.76381L21.3038 7.76385L20.6521 11.3697L19.1419 11.0968L19.5763 8.6934ZM2.60513 11.8598L5.41353 12.3376L5.1561 13.8505L2.32415 13.3687L3.53175 15.4965L2.19707 16.254L0.370968 13.0363L0.369261 13.036L0.369617 13.0339L0.194744 12.7258C0.0137573 12.4069 0.0836945 12.0045 0.361671 11.7654L0.623848 11.5398L0.626691 11.5231L0.640531 11.5255L3.45011 9.10849L4.45095 10.2719L2.60513 11.8598ZM6.84724 5.31605L8.71625 7.39819L7.57422 8.42332L5.66197 6.29301L5.21682 8.71915L3.70737 8.4422L4.3665 4.84982L4.36348 4.84646L4.36783 4.84256L4.43462 4.47854C4.4993 4.12603 4.79938 3.86525 5.15746 3.85036L5.48852 3.83659L5.50552 3.82133L5.51811 3.83536L9.26944 3.67933L9.33322 5.21265L6.84724 5.31605Z" fill="black" />
+        <path fillRule="evenodd" clipRule="evenodd" d="M14.402 0.492068C14.1568 0.200153 13.738 0.125161 13.4067 0.313837L13.0967 0.490347L13.0926 0.489595L13.092 0.493029L9.75878 2.39129L10.5403 3.76361L12.7171 2.52392L12.1815 5.42557L13.7345 5.71223L14.2732 2.79391L15.8774 4.7034L17.0866 3.68754L14.6448 0.781041L14.6457 0.776261L14.6399 0.775191L14.402 0.492068ZM9.70257 25C10.5056 25 11.1567 24.3654 11.1567 23.5826C11.1567 22.7998 10.5056 22.1653 9.70257 22.1653C8.8995 22.1653 8.24847 22.7998 8.24847 23.5826C8.24847 24.3654 8.8995 25 9.70257 25ZM11.6579 17.2218L11.1341 20.1425L9.57964 19.8637L10.1002 16.9611L7.9302 18.2072L7.14378 16.8376L10.4643 14.9309L10.465 14.9269L10.4698 14.9277L10.7813 14.7489C11.1114 14.5593 11.5298 14.6322 11.7763 14.9223L12.0166 15.2051L12.0195 15.2057L12.0191 15.208L14.4821 18.1065L13.2787 19.1291L11.6579 17.2218ZM12.9159 10.2489L12.3924 13.168L10.838 12.8892L11.3563 9.99873L9.1943 11.234L8.41086 9.86274L11.7197 7.97226L11.7233 7.95209L11.7475 7.95641L12.0493 7.78397C12.3793 7.59541 12.7969 7.66875 13.0429 7.95851L13.2735 8.23008L13.2778 8.23085L13.2772 8.2344L15.7396 11.1345L14.5358 12.1566L12.9159 10.2489ZM2.72286 21.7738L2.71439 21.7796L2.70749 21.7694L2.38579 21.677C2.04503 21.5792 1.81944 21.2559 1.84514 20.9023L2.13142 16.9638L3.62643 17.0724L3.45232 19.4679L5.78088 17.8886L6.62225 19.1291L4.33831 20.6782L6.66893 21.3474L6.25521 22.7882L2.72286 21.7738ZM18.1706 23.4104L18.1707 23.4106L18.1704 23.4108L18.0653 23.7567C17.9639 24.0901 17.6462 24.3097 17.2985 24.2866L16.9699 24.2648L16.9492 24.2794L16.9373 24.2626L13.2768 24.0196L13.3761 22.5239L15.8157 22.6859L14.2249 20.4493L15.4464 19.5805L17.073 21.8673L17.7769 19.5518L19.211 19.9878L18.1706 23.4104ZM22.4612 15.2819L22.4648 15.2825L22.4641 15.2869L22.6388 15.6C22.8176 15.9202 22.7447 16.3222 22.465 16.5593L22.2007 16.7833L22.1989 16.7939L22.1901 16.7924L19.3462 19.2031L18.3538 18.0325L20.2248 16.4465L17.4224 15.9534L17.6884 14.442L20.5122 14.9389L19.328 12.8174L20.668 12.0694L22.4612 15.2819ZM19.5763 8.6934L17.0442 10.0559L16.317 8.70445L18.8048 7.36579L16.5187 6.44108L17.0942 5.01841L20.5611 6.42075L20.5766 6.41239L20.5867 6.43112L20.9011 6.55827C21.2379 6.69454 21.433 7.04845 21.3684 7.40606L21.3038 7.76371L21.3038 7.76381L21.3038 7.76385L20.6521 11.3697L19.1419 11.0968L19.5763 8.6934ZM2.60513 11.8598L5.41353 12.3376L5.1561 13.8505L2.32415 13.3687L3.53175 15.4965L2.19707 16.254L0.370968 13.0363L0.369261 13.036L0.369617 13.0339L0.194744 12.7258C0.0137573 12.4069 0.0836945 12.0045 0.361671 11.7654L0.623848 11.5398L0.626691 11.5231L0.640531 11.5255L3.45011 9.10849L4.45095 10.2719L2.60513 11.8598ZM6.84724 5.31605L8.71625 7.39819L7.57422 8.42332L5.66197 6.29301L5.21682 8.71915L3.70737 8.4422L4.3665 4.84982L4.36348 4.84646L4.36783 4.84256L4.43462 4.47854C4.4993 4.12603 4.79938 3.86525 5.15746 3.85036L5.48852 3.83659L5.50552 3.82133L5.51811 3.83536L9.26944 3.67933L9.33322 5.21265L6.84724 5.31605Z" fill="black" />
     </svg>)
 }
