+++ src/app/utils/svg/upArrow.tsx
@@ -1,0 +1,5 @@
+export const UpArrowSvg = () => {
+    return <svg style={{ background: "black", margin: 9, borderRadius: 4 }} xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20" fill="black" className="absolute right-0 top-0 mt-3 mr-3">
+        <path d="M3.33301 9.99998L4.50801 11.175L9.16634 6.52498V16.6666H10.833V6.52498L15.483 11.1833L16.6663 9.99998L9.99967 3.33331L3.33301 9.99998Z" fill="white" />
+    </svg>
+}
